Implement getItem as a specialized version of loadItems
`marian::createScorer()` calls `loadItems()` twice: once for looking for a special:model.yml entry, and once for the actual data. The first call is very expensive since it will decode & dequantize all entries just for a simple lookup of a text file. This change copies some of the logic from `loadItems()` into `getItem()` to only process that one entry that it searches for. The alternative, passing the actual items down through createScorer and all the other users of the output of `loadItems()` would be better, but requires a far more intrusive change that doesn't yield any additional performance benefit in the standard bergamot-translator use case.